updated recalculation of best tap pair utilities for person specific attributes for tour and trip mode choice
git-svn-id: http://216.243.97.229:9090/svn/projects/trunk/mtctm2@7241 ee1c2863-5742-f242-9f82-d2f51c100d09

+++ src/java/com/pb/mtctm2/abm/ctramp/TransitWalkAccessDMU.java
@@ -149,7 +149,7 @@ public class TransitWalkAccessDMU
     	return personType;
     }
     
-    public void setValueOfTime(int valueOfTime) {
+    public void setValueOfTime(float valueOfTime) {
     	this.valueOfTime = valueOfTime;
     }
